+++ lib/infrastructures/services/cache/theme_mode_cache_impl.dart
@@ -8,7 +8,8 @@ import 'package:screen_pal/core/services/cache/theme_mode_cache.dart';
 class ThemeModeCacheImpl implements ThemeModeCache {
   final SharedPreferences _pref;
 
-  const ThemeModeCacheImpl(SharedPreferences preferences) : _pref = preferences;
+  const ThemeModeCacheImpl(SharedPreferences sharedPreferences)
+      : _pref = sharedPreferences;
 
   @protected
   @visibleForTesting