+++ test/helpers/mocks/external.dart
@@ -1,0 +1,6 @@
+import 'package:dio/dio.dart';
+import 'package:mockito/annotations.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+@GenerateMocks([Dio, SharedPreferences])
+void main() {}