+++ test/helpers/mocks/internal.dart
@@ -1,0 +1,11 @@
+import 'package:mockito/annotations.dart';
+
+import 'package:screen_pal/core/services.dart';
+
+@GenerateMocks([
+  ThemeModeCache,
+  ExtrasRepo,
+  MoviesRepo,
+  TvShowsRepo,
+])
+void main() {}