+++ test/helpers/mocks/services.dart
@@ -1,9 +1,0 @@
-import 'package:dio/dio.dart';
-import 'package:mockito/annotations.dart';
-import 'package:screen_pal/infrastructures/cache/theme_mode_cache.dart';
-
-@GenerateMocks([
-  Dio,
-  ThemeModeCache,
-])
-void main() {}